requirements.txt
fastapi==0.100.0
openai==0.27.7
pydantic==1.10.8
PyJWT==2.8.0
PyPDF2==3.0.1
python_bcrypt==0.3.2
Requests==2.31.0
uvicorn==0.23.1
waitress==2.1.2
